assignment-6/assignment.py
"""
Assignment 6: Advanced RAG Techniques
Student Name: [Your Name Here]
Date: [Date]

Instructions:
1. Implement different chunking strategies
2. Add metadata filtering capabilities
3. Create evaluation metrics for RAG performance
4. Compare and optimize different configurations
5. Generate comprehensive benchmark reports
"""

import os
import json
import time
from typing import List, Dict, Any, Tuple
from dataclasses import dataclass
from dotenv import load_dotenv
import pandas as pd
import numpy as np
from tqdm import tqdm

# TODO: Import necessary libraries
# Hint: You'll need all the imports from assignment 5 plus evaluation libraries

load_dotenv()

@dataclass
class ChunkingConfig:
    """Configuration for different chunking strategies"""
    strategy: str
    chunk_size: int
    overlap: int
    additional_params: Dict[str, Any] = None

@dataclass
class EvaluationResult:
    """Results from evaluating RAG performance"""
    retrieval_metrics: Dict[str, float]
    response_quality: Dict[str, float]
    performance_metrics: Dict[str, float]
    config: ChunkingConfig

class AdvancedRAG:
    """Advanced RAG system with optimization and evaluation capabilities"""
    
    def __init__(self, documents_path: str = "documents"):
        """Initialize the advanced RAG system"""
        self.documents_path = documents_path
        
        # TODO: Initialize components (similar to assignment 5)
        # Add support for multiple chunking strategies
        
        # Evaluation components
        self.test_questions = []
        self.ground_truth = {}
        self.evaluation_results = []
        
        # Load test data
        self._load_test_data()
    
    def _load_test_data(self):
        """Load test questions and ground truth answers"""
        # TODO: Load test questions from test_data/test_questions.json
        # TODO: Load ground truth from test_data/ground_truth.json
        
        pass
    
    def implement_chunking_strategy(self, strategy: str, config: ChunkingConfig):
        """
        Implement different chunking strategies
        
        Args:
            strategy: Type of chunking ('fixed', 'semantic', 'hierarchical')
            config: Configuration for the chunking strategy
        """
        
        if strategy == 'fixed':
            # TODO: Implement fixed-size chunking
            # Use RecursiveCharacterTextSplitter with specified chunk_size and overlap
            pass
            
        elif strategy == 'semantic':
            # TODO: Implement semantic chunking
            # Split by sentences or paragraphs, preserving meaning
            # Hint: Use nltk sentence tokenizer or similar
            pass
            
        elif strategy == 'hierarchical':
            # TODO: Implement hierarchical chunking
            # Create parent-child relationships between chunks
            pass
            
        else:
            raise ValueError(f"Unknown chunking strategy: {strategy}")
    
    def add_metadata_filtering(self, documents: List[Any]) -> List[Any]:
        """
        Add metadata to documents for better filtering
        
        Args:
            documents: List of loaded documents
            
        Returns:
            Documents with enhanced metadata
        """
        # TODO: Add metadata to documents
        # Include: source file, section, topic, difficulty level, etc.
        # Example metadata:
        # {
        #     'source': 'filename.txt',
        #     'section': 'Introduction',
        #     'topic': 'AI Basics',
        #     'difficulty': 'beginner'
        # }
        
        pass
    
    def implement_hybrid_search(self, query: str, k: int = 5) -> List[Any]:
        """
        Implement hybrid search combining BM25 and vector search
        
        Args:
            query: Search query
            k: Number of results to return
            
        Returns:
            Combined and re-ranked search results
        """
        # TODO: Implement BM25 search
        # TODO: Implement vector search
        # TODO: Combine and re-rank results
        # Hint: Use rank_bm25 library for BM25 implementation
        
        pass
    
    def evaluate_retrieval(self, query: str, retrieved_docs: List[Any], 
                          relevant_docs: List[str]) -> Dict[str, float]:
        """
        Evaluate retrieval performance
        
        Args:
            query: The search query
            retrieved_docs: Documents retrieved by the system
            relevant_docs: Ground truth relevant documents
            
        Returns:
            Dictionary with retrieval metrics
        """
        # TODO: Calculate retrieval metrics
        # - Precision@K: fraction of retrieved docs that are relevant
        # - Recall@K: fraction of relevant docs that are retrieved
        # - F1@K: harmonic mean of precision and recall
        
        metrics = {
            'precision_at_5': 0.0,
            'recall_at_5': 0.0,
            'f1_at_5': 0.0,
            'mrr': 0.0,  # Mean Reciprocal Rank
        }
        
        # TODO: Implement metric calculations
        
        return metrics
    
    def evaluate_response_quality(self, question: str, generated_answer: str, 
                                ground_truth_answer: str) -> Dict[str, float]:
        """
        Evaluate the quality of generated responses
        
        Args:
            question: The question asked
            generated_answer: Answer generated by RAG system
            ground_truth_answer: Expected correct answer
            
        Returns:
            Dictionary with response quality metrics
        """
        # TODO: Implement response quality evaluation
        # - Semantic similarity between generated and ground truth answers
        # - Factual accuracy (if ground truth is available)
        # - Completeness of the answer
        
        metrics = {
            'semantic_similarity': 0.0,
            'factual_accuracy': 0.0,
            'completeness': 0.0,
            'faithfulness': 0.0,  # How well answer reflects source material
        }
        
        # TODO: Use sentence transformers or similar for semantic similarity
        
        return metrics
    
    def measure_performance(self, query: str) -> Dict[str, float]:
        """
        Measure system performance metrics
        
        Args:
            query: Test query
            
        Returns:
            Performance metrics
        """
        start_time = time.time()
        
        # TODO: Execute query and measure performance
        
        end_time = time.time()
        
        metrics = {
            'latency': end_time - start_time,
            'tokens_used': 0,  # TODO: Track token usage
            'memory_usage': 0,  # TODO: Track memory if needed
        }
        
        return metrics
    
    def run_comprehensive_evaluation(self, configs: List[ChunkingConfig]) -> List[EvaluationResult]:
        """
        Run comprehensive evaluation across different configurations
        
        Args:
            configs: List of chunking configurations to test
            
        Returns:
            List of evaluation results for each configuration
        """
        results = []
        
        for config in tqdm(configs, desc="Testing configurations"):
            print(f"\nEvaluating config: {config.strategy} - {config.chunk_size}")
            
            # TODO: Set up RAG system with this configuration
            
            # Evaluate on test questions
            retrieval_scores = []
            quality_scores = []
            performance_scores = []
            
            for question in self.test_questions[:10]:  # Limit for testing
                # TODO: Run evaluation for this question
                # TODO: Collect metrics
                pass
            
            # TODO: Aggregate results
            result = EvaluationResult(
                retrieval_metrics={},
                response_quality={},
                performance_metrics={},
                config=config
            )
            
            results.append(result)
        
        return results
    
    def generate_benchmark_report(self, results: List[EvaluationResult]) -> str:
        """
        Generate comprehensive benchmark report
        
        Args:
            results: Evaluation results from different configurations
            
        Returns:
            Formatted report string
        """
        report = []
        report.append("="*60)
        report.append("RAG SYSTEM BENCHMARK REPORT")
        report.append("="*60)
        
        # TODO: Create detailed report with:
        # - Summary of configurations tested
        # - Performance comparison table
        # - Best performing configurations
        # - Recommendations for optimization
        
        # Convert to DataFrame for easy analysis
        df_data = []
        for result in results:
            row = {
                'strategy': result.config.strategy,
                'chunk_size': result.config.chunk_size,
                'overlap': result.config.overlap,
                **result.retrieval_metrics,
                **result.response_quality,
                **result.performance_metrics
            }
            df_data.append(row)
        
        df = pd.DataFrame(df_data)
        
        # TODO: Add analysis and recommendations
        
        return "\n".join(report)
    
    def save_results(self, results: List[EvaluationResult], filename: str = "benchmark_results.json"):
        """Save evaluation results to file"""
        # TODO: Serialize results and save to file
        pass

def create_test_data():
    """Create sample test data if it doesn't exist"""
    
    test_questions = [
        "What is artificial intelligence?",
        "How does machine learning work?",
        "What are the applications of Python in AI?",
        "Compare supervised and unsupervised learning",
        "What are the challenges in AI development?",
        "Explain neural networks in simple terms",
        "What is the difference between AI and machine learning?",
        "How do you evaluate a machine learning model?",
        "What are the ethical considerations in AI?",
        "What tools are used for AI development in Python?"
    ]
    
    # TODO: Create ground truth answers for evaluation
    ground_truth = {
        "What is artificial intelligence?": {
            "answer": "AI is a branch of computer science that aims to create intelligent machines...",
            "relevant_chunks": ["ai_basics.txt_chunk_1", "ai_basics.txt_chunk_2"]
        },
        # TODO: Add more ground truth data
    }
    
    # Create directories and save test data
    os.makedirs("test_data", exist_ok=True)
    
    with open("test_data/test_questions.json", "w") as f:
        json.dump(test_questions, f, indent=2)
    
    with open("test_data/ground_truth.json", "w") as f:
        json.dump(ground_truth, f, indent=2)

def main():
    """Main function to run the advanced RAG evaluation"""
    
    # Create test data if needed
    if not os.path.exists("test_data"):
        create_test_data()
    
    # Initialize advanced RAG system
    rag = AdvancedRAG()
    
    # Define configurations to test
    configs = [
        ChunkingConfig("fixed", 500, 100),
        ChunkingConfig("fixed", 1000, 200),
        ChunkingConfig("fixed", 1500, 300),
        ChunkingConfig("semantic", 1000, 0),
        ChunkingConfig("hierarchical", 2000, 0, {"levels": 2}),
    ]
    
    print("Starting comprehensive RAG evaluation...")
    
    # Run evaluation
    results = rag.run_comprehensive_evaluation(configs)
    
    # Generate and display report
    report = rag.generate_benchmark_report(results)
    print(report)
    
    # Save results
    rag.save_results(results)
    
    print("\nEvaluation complete! Check benchmark_results.json for detailed results.")

if __name__ == "__main__":
    main() 