assignment-6/requirements.txt
# Core LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Vector databases and search
chromadb==0.4.20
rank-bm25==0.2.2

# Environment and utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# Data processing and analysis
numpy==1.24.3
pandas==2.0.3

# Evaluation and metrics
scikit-learn==1.3.0

# Text processing
nltk==3.8.1
spacy==3.7.2

# Performance and timing
tqdm==4.66.1

# Visualization (optional)
matplotlib==3.8.0
seaborn==0.13.0 